+++ stack.py
@@ -10,22 +10,6 @@ class Node(object):
         self.data = data
         self.next_node = next_node
 
-    def set_next(self, next_node):
-        """Method designed to set next_node of Node object."""
-        self.next_node = next_node
-
-    def set_data(self, data):
-        """Method designed to set data of Node object."""
-        self.data = data
-
-    def get_next_node(self):
-        """Method designed to return next_node of Node object."""
-        return self.next_node
-
-    def get_data(self):
-        """Method designed to return data of Node object."""
-        return self.data
-
 
 class Stack(object):
     """Class designed to create stack objects."""
